+++ openfeature-provider-intellitoggle/lib/openfeature_provider_intellitoggle.dart
@@ -27,6 +27,8 @@ export 'src/utils.dart'
 // Re-export OpenFeature types
 export 'package:openfeature_dart_server_sdk/open_feature_api.dart';
 export 'package:openfeature_dart_server_sdk/evaluation_context.dart';
+export 'package:openfeature_dart_server_sdk/hooks.dart';
+export 'package:openfeature_dart_server_sdk/multi_provider.dart';
 export 'hooks/intellitoggle_telemetry_hook.dart';
 export 'package:openfeature_dart_server_sdk/feature_provider.dart'
     hide InMemoryProvider;